requirements.txt
locust>=2.17.0
pandas>=2.0.0
matplotlib>=3.7.0
psutil>=5.9.0
pyyaml>=6.0
numpy>=1.24.0
scipy>=1.16.0
boto3>=1.26.0
python-dotenv>=1.0.0